// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

/** A Field Group registered by ACF */
export type AcfFieldGroup = {
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
};
// ...
export type Page_Informacoesdecontato = AcfFieldGroup & {
  __typename?: 'Page_Informacoesdecontato';
  coEmail?: Maybe<Scalars['String']>;
  coEndereco?: Maybe<Scalars['String']>;
  coTelefone?: Maybe<Scalars['String']>;
  coWhatsapp?: Maybe<Scalars['String']>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  linkFacebook?: Maybe<Scalars['String']>;
  linkInstagram?: Maybe<Scalars['String']>;
  linkYoutube?: Maybe<Scalars['String']>;
};